requirements.txt
pygame-ce==2.5.6
kivy>=2.3,<3.0
numpy>=2.0,<3.0
pytest>=8.0,<9.0
